Get old and new versions building
