+++ src/Main.lhs
@@ -15,7 +15,7 @@ Path settings auto-generated by Cabal:
 > import AbsSyn
 > import Grammar
 > import PrettyGrammar
-> import Parser
+> import Parser.Oracle
 > import First
 > import LALR
 > import ProduceCode (produceParser)
@@ -76,7 +76,7 @@ Open the file.
 
 Parse, using bootstrapping parser.
 
->       case runP ourParser (PS file 1 Nothing) of {
+>       case runFromStartP ourParser file 1 of {
 >               Left err -> die (fl_name ++ ':' : err);
 >               Right abssyn@(AbsSyn hd _ _ tl) ->
