+++ src/Parser/Oracle.hs
@@ -9,6 +9,7 @@ import Control.Applicative
 import Control.Monad (when)
 import Data.Maybe (isJust)
 import ParseMonad
+import ParseMonad.Oracle
 import AbsSyn
 import Lexer
